pyproject.toml
[tool.poetry]
name = "lexz"
version = "0.1.0"
description = "Time traveler for obfuscation or another purpose"
authors = ["krypton-byte <rosid6434@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
include = ["assets/output.svg"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
